fix(test) after

+++ contracts/Users/Interfaces/ICDAOAdmins.sol
@@ -86,7 +86,7 @@ interface ICDAOAdmins {
     
     function checkEligibility(address _user) external view returns (bool);
 
-    function checkEligibility(address _user, uint256 _snap, address _batchId) external returns (bool);
+    function checkEligibility(address _user, uint _snap, address _batchId) external returns (bool);
 
     function createCecaTokenForBatch(address _batch, uint _index) external;
